feat: add SSL connection support for the MySQL database
This is for the new AWS database which requires using SSL connections. It also runs on a different port than default. To make it easy to switch between database servers without having to deploy different code, this change is backwards compatible with the old database and the old permissions file that only had 4 lines in it.